eccc/file_server/requirements.txt
google-cloud-storage==1.35.0
flask==1.1.2
gunicorn==20.0.4
pip==20.3.3
flask==1.1.2
python-dotenv==0.15.0
smart-open==4.1.2
